requirements.txt
absl-py==0.11.0
astor==0.8.1
beautifulsoup4==4.7.1
cached-property==1.5.2
cycler==0.10.0
Django==2.2.6
gast==0.4.0
grpcio==1.35.0
h5py==2.10.0
importlib-metadata==3.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.0
mock==4.0.3
numpy==1.20.1
opencv-python==4.1.1.26
pandas==1.0.3
Pillow==7.2.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.11.1
scikit-learn==0.20.3
scipy==1.6.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.0
